Don't assume pid_t is int
Add (int) casts when passing a pid_t as a "%d" value etc

+++ source3/torture/test_messaging_send_all.c
@@ -270,7 +270,7 @@ bool run_messaging_send_all(int dummy)
 		} while ((child == -1) && (errno == EINTR));
 
 		if (child != children[i]) {
-			printf("waitpid(%d) failed\n", children[i]);
+			printf("waitpid(%d) failed\n", (int)children[i]);
 			return false;
 		}
 	}
